+++ CHANGELOG.md
@@ -1,1 +1,2 @@
 * Fixes an auth issue with `https.onCall` functions when runtime options (`runWith`) are specified (#2059).
+* Fixed an issue with the Firestore emulator where ternary operators were not allowed.